pyproject.toml
[project]
name = "hexlet-code"
version = "0.1.0"
description = "Compares two configuration files and shows a difference"
authors = [
	{name = 'farguz', email = 'aaaaa@bbb.cc'}
]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "pyyaml>=6.0.2",
]

[dependency-groups]
dev = [
    "pytest>=8.3.4",
    "pytest-cov>=6.0.0",
    "ruff>=0.9.7",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["gendiff"]

[project.scripts]
gendiff = "gendiff.scripts.gendiff:main"
